requirements.txt
# FastAPI and web framework dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sse-starlette>=1.8.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Claude Agent SDK
anyio>=4.0.0
claude-agent-sdk

# Database (for context data access if needed)
pymysql>=1.0.0
tqdm

# CLI tool dependencies
prompt-toolkit>=3.0.43    # Interactive REPL
rich>=13.7.0              # Terminal formatting